[Feature #154953693] Allow creation of users via REST API.
This feauture will make it possible to create new users via the API. External apps might need to register accounts on behalf of their users without having to send them over to the wger instance. A workflow like this is suggested:

Apps register as a regular user to get an API key
By default, users aren't allowed to create new accounts over the API, this has to be explicitly allowed
Users created with this key are marked somehow so it is later possible to identify them, you should probably think about adding a new OneToOne table to the user table "metadata" or such.

Simple idea for a basic protection against being spammed to death with new users:

REST API consumers that want to create users need a special right that has to be given on demand.
Users generated could have a special flag so that it's possible to see who created them

Note that it's probably easier and faster to add a new command for this (instead of a new element in the GUI). Something like add-user-rest-api and list-user-rest-apiwould be enough.