[Feature #154953697] Make translations of exercises and ingredients easier.
This was archieved by adding a filter option to filter the displayed ingredients by preferred language.

Below is the description of the task:
Every ingredient and exercise has a field indicating the language, which is automatically set as the user's current language when adding them.
The current list is then filtered based on this.

The user should be able to select which languages to show. They should not be shown only their current language.